Добавил функциональную возможность работы с курсами(управление ими через контроллер и сервис)

+++ src/main/java/ru/tihomirov/university/service/CourseServiceImpl.java
@@ -2,11 +2,11 @@ package ru.tihomirov.university.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.tihomirov.university.exception.EntityNotFoundException;
 import ru.tihomirov.university.model.Course;
 import ru.tihomirov.university.repository.CourseRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -20,8 +20,9 @@ public class CourseServiceImpl implements CourseService {
     }
 
     @Override
-    public Optional<Course> getById(Long id) {
-        return courseRepository.findById(id);
+    public Course getById(Long id) {
+        return courseRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Course not found with id: " + id));
     }
 
     @Override
@@ -31,6 +32,9 @@ public class CourseServiceImpl implements CourseService {
 
     @Override
     public void deleteById(Long id) {
+        if (!courseRepository.existsById(id)) {
+            throw new EntityNotFoundException("Course not found with id: " + id);
+        }
         courseRepository.deleteById(id);
     }
 }
